+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add endpoints to show the watchers list for issues, tasks and user stories.
 - Add headers to allow threading for notification emails about changes to issues, tasks, user stories, and wiki pages. (thanks to [@brett](https://github.com/brettp)).
 - Add externall apps: now Taiga can integrate with hundreds of applications and service.
+- Add filter by email domain on register.
 - i18n.
   - Add polish (pl) translation.
   - Add portuguese (Brazil) (pt_BR) translation.